% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badwell.R
\name{badwell}
\alias{badwell}
\title{Removes a bad replicate, then calculates the Mean and standard deviation between replicates}
\usage{
badwell(df, well)
}
\arguments{
\item{df}{A dataframe.}

\item{well}{A column header}
}
\value{
A dataframe.
}
\description{
Removes a bad replicate, then calculates the Mean and standard deviation between replicates
}
