% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platemeanstd.R
\name{platemeanstd}
\alias{platemeanstd}
\title{Calculating the Mean and standard deviation between replicates}
\usage{
platemeanstd(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Calculating the Mean and standard deviation between replicates
}
